requirements.txt
spotipy
python-dotenv
sentence-transformers
pylast
PySimpleGUI
PyQt5